feat : projectile, damage
투사체 스크립트 작성 및 정리

+++ Assets/02_Scripts/Interfaces/IDamageable.cs
@@ -1,4 +1,4 @@
 ﻿public interface IDamageable
 {
-    public bool TakeDamage(int damage);
+    public bool TakeDamage(float damage);
 }
